% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dsGeospatial-package.R
\docType{package}
\name{dsGeospatial-package}
\alias{dsGeospatial}
\alias{dsGeospatial-package}
\title{dsGeospatial: Privacy-Preserving Geospatial Analysis in DataSHIELD}
\description{
Provides privacy-preserving methods for the visualisation and analysis of linked health and environmental data using the DataSHIELD framework. Individual-level and household-level data remain within the data-holding organisations, while only disclosure- controlled aggregate results are returned to the analyst. The package is being developed as part of the GROVE project within the DARE UK programme.
}
\author{
\strong{Maintainer}: Theertha Kariyathan \email{Theertha.Kariyathan@liverpool.ac.uk}

Authors:
\itemize{
  \item Theertha Kariyathan \email{Theertha.Kariyathan@liverpool.ac.uk}
}

}
\keyword{internal}
